Refine observation detail presentation
