+++ web/src/features/traces/components/IOPreview/IOPreview.tsx
@@ -14,6 +14,7 @@ import { type ChatMLParserResult } from "../../hooks/useChatMLParser";
 import { Button } from "@/src/components/ui/button";
 import { ActionButton } from "@/src/components/ActionButton";
 import { BookOpen, X } from "lucide-react";
+import type { ObservationStatusMessage } from "./components/statusMessagePresentation";
 
 export type { ViewMode };
 export type IOPreviewContentMode = "all" | "conversation";
@@ -53,7 +54,7 @@ export interface ExpansionStateProps {
 export interface IOPreviewProps extends ExpansionStateProps {
   input?: Prisma.JsonValue;
   output?: Prisma.JsonValue;
-  statusMessage?: string;
+  status?: ObservationStatusMessage;
   metadata?: Prisma.JsonValue;
   outputCorrection?: ScoreDomain;
   // Pre-parsed data (optional, from useParsedObservation hook for performance)
@@ -105,7 +106,7 @@ export interface IOPreviewProps extends ExpansionStateProps {
 export function IOPreview({
   input,
   output,
-  statusMessage,
+  status,
   outputCorrection,
   metadata,
   parsedInput,
@@ -180,7 +181,7 @@ export function IOPreview({
   const sharedProps = {
     input,
     output,
-    statusMessage,
+    status,
     outputCorrection,
     metadata,
     parsedInput,
@@ -242,7 +243,7 @@ export function IOPreview({
         <IOPreviewJSON
           input={input}
           output={output}
-          statusMessage={statusMessage}
+          status={status}
           metadata={metadata}
           parsedInput={parsedInput}
           parsedOutput={parsedOutput}
@@ -269,7 +270,7 @@ export function IOPreview({
         <IOPreviewJSONSimple
           input={input}
           output={output}
-          statusMessage={statusMessage}
+          status={status}
           metadata={metadata}
           outputCorrection={outputCorrection}
           parsedInput={parsedInput}
