+++ web/src/features/traces/components/IOPreview/IOPreviewPretty.tsx
@@ -16,7 +16,8 @@ import {
 } from "./IOPreview";
 import { CorrectedOutputField } from "./components/CorrectedOutputField";
 import { isOnlyJsonMessage } from "../../fns/chatMessageUtils";
-import { ErrorSection } from "./components/ErrorSection";
+import { StatusMessageSection } from "./components/StatusMessageSection";
+import type { ObservationStatusMessage } from "./components/statusMessagePresentation";
 
 interface JsonInputOutputViewProps {
   parsedInput: unknown;
@@ -87,7 +88,7 @@ function JsonInputOutputView({
 export interface IOPreviewPrettyProps extends ExpansionStateProps {
   input?: Prisma.JsonValue;
   output?: Prisma.JsonValue;
-  statusMessage?: string;
+  status?: ObservationStatusMessage;
   metadata?: Prisma.JsonValue;
   outputCorrection?: ScoreDomain;
   // Pre-parsed data (optional, from useParsedObservation hook for performance)
@@ -128,7 +129,7 @@ export interface IOPreviewPrettyProps extends ExpansionStateProps {
 export function IOPreviewPretty({
   input,
   output,
-  statusMessage,
+  status,
   metadata,
   outputCorrection,
   parsedInput: preParsedInput,
@@ -275,9 +276,7 @@ export function IOPreviewPretty({
 
   return (
     <div>
-      {showData && statusMessage ? (
-        <ErrorSection message={statusMessage} />
-      ) : null}
+      {showData && status ? <StatusMessageSection status={status} /> : null}
 
       {showData ? (
         <SectionToolDefinitions
